fixed paths and tests and storybook
