+++ package.json
@@ -68,6 +68,7 @@
     "@babel/core": "^7.9.0",
     "@babel/node": "^7.8.7",
     "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-private-methods": "7.12.13",
     "@babel/polyfill": "^7.8.7",
     "@babel/preset-env": "7.9.5",
     "@babel/preset-flow": "7.9.0",
