+++ packages/graphiql-react/package.json
@@ -33,13 +33,14 @@
     "react"
   ],
   "dependencies": {
-    "@emotion/core": "11.0.0",
-    "theme-ui": "0.3.5"
+    "@emotion/core": "10.0.28",
+    "csstype": "2.6.6",
+    "theme-ui": "0.3.1"
   },
   "devDependencies": {
-    "@storybook/react": "6.1.17",
     "@storybook/addon-essentials": "6.1.17",
     "@storybook/addon-storyshots": "6.0.22",
+    "@storybook/react": "6.1.17",
     "jest": "26.6.3"
   },
   "peerDependencies": {
