+++ packages/graphiql-react/src/components/Type.stories.tsx
@@ -11,7 +11,7 @@
 import { jsx } from 'theme-ui';
 import List, { ListRow } from './List';
 import { SectionHeader, Explainer } from './Type';
-import { layout } from './themes/decorators';
+import { layout } from '../themes/decorators';
 
 export default { title: 'Type', decorators: [layout] };
