+++ packages/graphiql-react/src/components/__tests__/stories.spec.js
@@ -9,6 +9,4 @@
 import initStoryshots from '@storybook/addon-storyshots';
 import path from 'path';
 
-initStoryshots({
-  configPath: path.resolve(__dirname, '../../../.storybook'),
-});
+initStoryshots();
