+++ packages/graphiql-react/src/themes/README.md
