+++ packages/graphiql-react/src/themes/default/Layout.tsx
