+++ packages/graphiql-react/src/themes/default/index.ts
