+++ packages/graphiql-react/src/themes/types.ts
@@ -6,7 +6,7 @@
  *
  */
 
-import { Maybe, ReactNodeLike } from '../../../types';
+import { Maybe, ReactNodeLike } from '../types';
 
 export type Color = string;
 export type Colors = {
